Add the last filelist for data of 2017 skim
The date is different from the other filelists with the date 2019-05-15
because one of the jobs made problems and the easiest solution was
to just resubmit the task, the date is different so we can differentiate
it from the 2019-05-15 skim, to be sure that this filelist is complete.
Now the 2017 data is fully skimmed.