Calculate the Covariance Matrix for the helical approach
The errors are propagated using Jacobi Matrices.
A lot of new variables have been introduced to store the covariance
matrices both for the IP with reconstructed and refitted PV.
The covariance matrix of the parameter O' has not been considered yet
but implementation is trivial (currently the covariance matrix for O' is
just 0).