Change the name of variables for more consistant and shorter names
The Variables that are relevant to the Higgs CP final state analysis are
now changed to follow the naming convention more closely. These changes
mostly concern variable with extra information in their name such as
recoPhiStarCP_refitPV becomes recoPhiStarCPrPV.
The Changes are:
_helical -> Hel
_refitPV -> rPV
_rho -> Rho
_method -> Meth
The Order in which these abbreviations are used is
HelrPVBS (i.e. you will not find a variable "recoPhiStarCPrPVHel" but rather
"recoPhiStarCPHelrPV)
The numbering is now always at the end of the name with an underscore
before the number itself (IP1_refitPV -> IPrPV_1).

The other big change is that vectors are saved as RMPoints instead of
saving each component individually.

Since a few other files use the same variables of the product as the
RecoTauCPPRoducer, the other listed files were also changed. But they
were only changed to work with this commit, not to make every variable
in those files consistent with this so there might be some variables of
the old naming convention. The Function names in CPQuantities were
adjusted too.
The RefitVertexSelector now saves vectors such as the refitPV as a
RMPoint but some variables may still store the individual components.
This will be fixed soon.