Add the QuantileMappingProducer and rename related quantities
The QuantileMappingProducer uses Z->mumu events for calibration by
subtracting the generator level IP. Since it was shown that this works
well, the Quantities with the calib prefix.
Therefore the quantities calculated by the IsomorphicMappingProducer now
have the isomap prefix to distinguish them.
Lastly, the isomorphic mapping is only done for the refitted PV and
refitted PV with BS constraint. Quantile Mapping is only done for the
refitted PV with BS constraint, thus the unused quantities are removed.